+++ app/src/androidTest/java/com/ruangaldo/gubake/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.kertasguntingbatu
+package com.gubake.gubakegames
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
