+++ app/src/main/java/com/ruangaldo/gubake/playGame/MainGameComputer.kt
@@ -1,4 +1,4 @@
-package com.example.kertasguntingbatu.playGame
+package com.gubake.gubakegames.playGame
 
 import android.content.Intent
 import android.os.Bundle
@@ -15,27 +15,21 @@ import androidx.appcompat.app.AppCompatActivity
 import androidx.constraintlayout.widget.ConstraintLayout
 import androidx.core.view.isVisible
 import com.bumptech.glide.Glide
-import com.example.kertasguntingbatu.MainMenu
-import com.example.kertasguntingbatu.R
+import com.gubake.gubakegames.MainMenu
+import com.gubake.gubakegames.R
 import com.google.android.material.snackbar.Snackbar
 import kotlinx.android.synthetic.main.activity_maingame.*
 
 
-class MainGamePlayer : AppCompatActivity(), IControllerNya {
+class MainGameComputer : AppCompatActivity(), IControllerNya {
     private val imgLogo = "https://i.ibb.co/HC5ZPgD/splash-screen1.png"
     private var dataPlayer1 = ""
-    private var dataPlayer2 = ""
     private val layoutImage: ConstraintLayout by lazy { findViewById(R.id.activity_maingame) }
     private val resetFun by lazy {
         findViewById<ImageView>(R.id.imageBattle)
     }
+    private val intentDialog by lazy { Intent(this, MainMenu::class.java) }
     private var enemyNya: ImageButton? = null
-    private val textName by lazy {
-        mutableListOf(
-                findViewById<TextView>(R.id.player1),
-                findViewById(R.id.player2)
-        )
-    }
     private val backgroundAll by lazy {
         mutableListOf(
                 findViewById<FrameLayout>(R.id.backgroundBatu), findViewById(R.id.backgroundScissors),
@@ -45,36 +39,38 @@ class MainGamePlayer : AppCompatActivity(), IControllerNya {
     }
     private val buttonAll by lazy {
         mutableListOf(
-                findViewById(R.id.batuPlayer),
-                findViewById(R.id.scissorsPlayer),
-                findViewById(R.id.paperPlayer),
-                findViewById(R.id.batuComp),
-                findViewById(R.id.scissorsComp),
-                findViewById(R.id.paperComp),
-                findViewById(R.id.refreshBut),
-                findViewById<ImageButton>(R.id.but_close)
+                findViewById(R.id.batuPlayer), findViewById(R.id.scissorsPlayer),
+                findViewById(R.id.paperPlayer), findViewById(R.id.batuComp), findViewById(R.id.scissorsComp),
+                findViewById(R.id.paperComp), findViewById(R.id.refreshBut), findViewById<ImageButton>(R.id.but_close)
         )
     }
-    private val intentDialog by lazy { Intent(this, MainMenu::class.java) }
+    private val textName by lazy {
+        mutableListOf(
+                findViewById<TextView>(R.id.player1),
+                findViewById(R.id.player2)
+        )
+    }
     private val controller = ControllerNya(this)
     private val randDuration = 1000L
+    private val animRand = 200L
+    private var randNum = 0
     private var name = mutableListOf<String>()
     private var namePlay: String = "Pemain 1"
+
     override fun onCreate(savedInstanceState: Bundle?) {
+        val animation: Animation = AnimationUtils.loadAnimation(this, R.anim.bounce)
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_maingame)
-        namePlay = intent.getStringExtra("dataName").toString()
-        name.add(namePlay)
-        val animation: Animation = AnimationUtils.loadAnimation(this, R.anim.bounce)
-        namePlay = intent.getStringExtra("dataName").toString()
         Glide.with(this)
                 .load(imgLogo).error(R.drawable.ic_logo_image)
                 .into(logoImageGame)
-        textName[0]?.text = namePlay
-        textName[1].text = getString(R.string.Player2)
         reset()
+        namePlay = intent.getStringExtra("dataName").toString()
+        name.add(namePlay)
+        textName[0]?.text = name[0]
+        textName[1]?.text = getString(R.string.computer)
         val butNya = mutableListOf(
-                buttonAll[0], buttonAll[1], buttonAll[2], buttonAll[3], buttonAll[4], buttonAll[5]
+                buttonAll[0], buttonAll[1], buttonAll[2],
         )
         butNya.forEachIndexed { _, imageButton ->
             imageButton.setOnClickListener {
@@ -94,135 +90,120 @@ class MainGamePlayer : AppCompatActivity(), IControllerNya {
                         }
                         dataPlayer1 = "gunting"
                     }
-                    buttonAll[2] -> {
+                    else -> {
                         if (!backgroundAll[2].isVisible) {
                             buttonAll[2].startAnimation(animation)
                             backgroundAll[2].visibility = View.VISIBLE
                         }
                         dataPlayer1 = "kertas"
                     }
-                    buttonAll[3] -> {
-                        if (!backgroundAll[3].isVisible) {
-                            buttonAll[3].startAnimation(animation)
-                            backgroundAll[3].visibility = View.VISIBLE
-                        }
-                        dataPlayer2 = "batu"
-                    }
-                    buttonAll[4] -> {
-                        if (!backgroundAll[4].isVisible) {
-                            buttonAll[4].startAnimation(animation)
-                            backgroundAll[4].visibility = View.VISIBLE
-                        }
-                        dataPlayer2 = "gunting"
-                    }
-                    else -> {
-                        if (!backgroundAll[5].isVisible) {
-                            buttonAll[5].startAnimation(animation)
-                            backgroundAll[5].visibility = View.VISIBLE
-                        }
-                        dataPlayer2 = "kertas"
-                    }
                 }
-                lockButtonPlay1()
-                if (dataPlayer1 != "" && dataPlayer2 != "") {
-                    Toast.makeText(this, "$namePlay Memilih $dataPlayer1", Toast.LENGTH_SHORT)
-                            .show()
-                    lockButtonPlay2()
-                    dataModel()
-                    resetFun.animate().alpha(0f).scaleX(0.5f).scaleY(0.5f).setDuration(500).start()
-                    buttonAll[7].animate().alpha(0f).scaleX(0.5f).scaleY(0.5f).setDuration(500)
-                            .start()
-                }
-                Log.i("MainGamePlayer", "$namePlay memilih $dataPlayer1 dan Player 2 memilih $dataPlayer2 ")
-
+                Toast.makeText(this, "$namePlay Memilih $dataPlayer1", Toast.LENGTH_SHORT).show()
+                Log.i("MainGameComputer", "$namePlay memilih $dataPlayer1")
+                lockButton()
+                animationRandLoop()
+                resetFun.animate().alpha(0f).scaleX(0.5f).scaleY(0.5f).setDuration(500).start()
+                buttonAll[7].animate().alpha(0f).scaleX(0.5f).scaleY(0.5f).setDuration(500).start()
             }
         }
         buttonAll[6].setOnClickListener {
             buttonAll[6].startAnimation(animation)
             reset()
-            unlockButton()
-            unlockButton2()
-            Log.i("MainGamePlayer", "playernya klik reset")
+            Log.i("MainGameComputer", "playernya klik reset")
         }
         buttonAll[7].setOnClickListener {
             buttonAll[7].startAnimation(animation)
             onBackPressed()
-            Log.i("MainGamePlayer", "playernya klik exit")
+            Log.i("MainGameComputer", "playernya klik keluar")
         }
+    }
+
+    //Animasi Random Pada Suitnya
+    private fun animationRandLoop() {
+        mutableListOf(backgroundAll[3], backgroundAll[4], backgroundAll[5])
+                .forEachIndexed { _, i ->
+                    Handler(Looper.getMainLooper()).postDelayed({
+                        if (i.isVisible) {
+                            i.visibility = View.INVISIBLE
+                            Log.i("MainGameComputer", "Hilang Sejenak Background nya Computer")
+                        }
+                    }, animRand)
+                }
+        Handler(Looper.getMainLooper()).postDelayed({
+            if (randNum <= 35) {
+                controller.compRand()
+                Log.i("MainGameComputer", "Perulangan Computer #${randNum}")
+            } else {
+                dataModel()
+                randNum = 0
+            }
+        }, animRand)
     }
 
     //Pemrosesan Data
     private fun dataModel() {
-        val dataMauPlayer = ModelNya(dataPlayer1, dataPlayer2, "vsPlayer")
-        Log.i("MainGamePlayer", "Proses Suit Pemain vs Pemain")
+        val dataMauPlayer = ModelNya(dataPlayer1, "", "vsCPU")
+        Log.i("MainGameComputer", "Proses Suit Computer vs Pemain")
         controller.setDataPlayer(dataMauPlayer)
         controller.chooseEnemy()
         controller.compareData()
     }
 
     //Button di lock ketika di proses
-    private fun lockButtonPlay1() {
+    private fun lockButton() {
         buttonAll[0].isEnabled = false
         buttonAll[1].isEnabled = false
         buttonAll[2].isEnabled = false
-        Log.i("MainGamePlayer", "di Lock yekkk ga bisa main kapok kon")
+        Log.i("MainGameComputer", "di Lock yekkk ga bisa main kapok kon")
     }
-
-    private fun lockButtonPlay2() {
-        buttonAll[3].isEnabled = false
-        buttonAll[4].isEnabled = false
-        buttonAll[5].isEnabled = false
-        Log.i("MainGamePlayer", "di Lock yekkk ga bisa main kapok kon player 2")
-    }
-
 
     //unlock button semua
     private fun unlockButton() {
         if (!buttonAll[0].isEnabled && !buttonAll[1].isEnabled && !buttonAll[2].isEnabled) {
             buttonAll[0].isEnabled = true
-            Log.i("MainGamePlayer", "Cieee bisa buka batu")
+            Log.i("MainGameComputer", "Cieee bisa buka batu")
             buttonAll[1].isEnabled = true
-            Log.i("MainGamePlayer", "Cieee bisa buka gunting")
+            Log.i("MainGameComputer", "Cieee bisa buka gunting")
             buttonAll[2].isEnabled = true
-            Log.i("MainGamePlayer", "Cieee bisa buka kertas")
-        }
-    }
-
-    private fun unlockButton2() {
-        if (!buttonAll[3].isEnabled && !buttonAll[4].isEnabled && !buttonAll[5].isEnabled) {
-            buttonAll[3].isEnabled = true
-            Log.i("MainGamePlayer", "Cieee bisa buka batu player 2")
-            buttonAll[4].isEnabled = true
-            Log.i("MainGamePlayer", "Cieee bisa buka gunting player 2")
-            buttonAll[5].isEnabled = true
-            Log.i("MainGamePlayer", "Cieee bisa buka kertas player 2")
+            Log.i("MainGameComputer", "Cieee bisa buka kertas")
         }
     }
 
     //Reset semua
     private fun reset() {
         mutableListOf(
-                backgroundAll[0],
-                backgroundAll[1],
-                backgroundAll[2],
-                backgroundAll[3],
-                backgroundAll[4],
-                backgroundAll[5],
+                backgroundAll[0], backgroundAll[1], backgroundAll[2], backgroundAll[3], backgroundAll[4], backgroundAll[5],
         ).forEachIndexed { _, i ->
             if (i.isVisible) {
                 i.visibility = View.INVISIBLE
-                Log.i("MainGamePlayer", "Background nya ilang semua")
+                Log.i("MainGameComputer", "Background nya ilang semua")
             }
         }
         resetFun.animate().alpha(1f).scaleX(1f).scaleY(1f).setDuration(randDuration).start()
         buttonAll[7].animate().alpha(1f).scaleX(1f).scaleY(1f).setDuration(randDuration).start()
         dataPlayer1 = ""
-        dataPlayer2 = ""
-
+        unlockButton()
     }
 
     //Animasi Random dari Computer
-    override fun randAnim(animResult: String) {}
+    override fun randAnim(animResult: String) {
+        when (animResult) {
+            "batu" -> {
+                backgroundAll[3].visibility = View.VISIBLE
+                Log.i("MainGameComputer", "Acak acak ini batu")
+            }
+            "gunting" -> {
+                backgroundAll[4].visibility = View.VISIBLE
+                Log.i("MainGameComputer", "Acak acak ini gunting")
+            }
+            "kertas" -> {
+                backgroundAll[5].visibility = View.VISIBLE
+                Log.i("MainGameComputer", "Acak acak ini kertas")
+            }
+        }
+        randNum++
+        animationRandLoop()
+    }
 
     //Hasil dari Randomnya computer
     override fun resultEnemy(resultEnemy: String) {
@@ -240,8 +221,7 @@ class MainGamePlayer : AppCompatActivity(), IControllerNya {
                 backgroundAll[5].visibility = View.VISIBLE
             }
         }
-        Log.e("MainGamePlayer", "pilihan Pemain 2 $resultEnemy")
-        Toast.makeText(this, "Pemain 2 Memilih $resultEnemy", Toast.LENGTH_SHORT).show()
+        Toast.makeText(this, "CPU Memilih $resultEnemy", Toast.LENGTH_SHORT).show()
     }
 
     //Menampilkan hasil dari kalah atau menang
@@ -252,13 +232,13 @@ class MainGamePlayer : AppCompatActivity(), IControllerNya {
                 winner = "$namePlay\nMENANG!"
             }
             "P2Win" -> {
-                winner = "Player 2\nMENANG!"
+                winner = "CPU\nMENANG!"
             }
             "Seri" -> {
                 winner = "SERI!"
             }
         }
-        Log.i("MainGamePlayer", "pemenangnya $winner")
+        Log.i("MainGameComputer", "pemenangnya $winner")
         Handler(Looper.getMainLooper()).postDelayed({
             val view = LayoutInflater.from(this).inflate(R.layout.activity_dialog, null, false)
             val dialogBuilder = AlertDialog.Builder(this)
@@ -271,26 +251,27 @@ class MainGamePlayer : AppCompatActivity(), IControllerNya {
             winnerInfo.text = winner
             playAgain.setOnClickListener {
                 reset()
-                unlockButton()
-                unlockButton2()
                 dialogD1.dismiss()
             }
             backMenu.setOnClickListener {
-
                 intentDialog.putExtra("dataName", name[0])
                 startActivity(intentDialog)
+
             }
-            dialogD1.show()
+            if (!isFinishing) {
+                dialogD1.show()
+            }
+
         }, 2 * randDuration
         )
     }
 
     override fun onBackPressed() {
-        val snackPlayer = Snackbar.make(layoutImage, "Apakah ingin keluar?", Snackbar.LENGTH_SHORT)
-        snackPlayer.setAction("Keluar") {
-            snackPlayer.dismiss()
+        val snackComp = Snackbar.make(layoutImage, "Apakah ingin keluar?", Snackbar.LENGTH_SHORT)
+        snackComp.setAction("Keluar") {
+            snackComp.dismiss()
             finish()
         }.show()
-        Log.i("MainGamePlayer", "Keluar")
+        Log.i("MainGameComputer", "Keluar")
     }
 }
