+++ app/src/main/java/com/ruangaldo/gubake/playGame/ModelNya.kt
@@ -1,4 +1,4 @@
-package com.example.kertasguntingbatu.playGame
+package com.gubake.gubakegames.playGame
 
 data class ModelNya(val dataPlayer1: String, val dataPlayer2:String, val modeGame:String)
